refactor : 멘토 - 페이지들 모두 content 컴포넌트를 이용해서 클라이언트 컴포넌트 분리했습니다
